chore: made a Event attendanceVerification file for clean code
